- Move code from PhotosController to PhotoService and FileSystemPhotoStorage. - Move filter logic from VehicleRepository to IQueryableExtensions. - Use chart.js and angular2-chartjs libraries to create chart. - Utilize UserSecrets library to store sensitive data, i.e. connection strings.
