+++ ClientApp/package.json
@@ -26,8 +26,10 @@
     "@fortawesome/fontawesome-free": "^5.9.0",
     "@nguniversal/module-map-ngfactory-loader": "6.0.0",
     "@sentry/browser": "^5.6.3",
+    "angular2-chartjs": "^0.5.1",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.3.1",
+    "chart.js": "^2.8.0",
     "core-js": "^2.5.4",
     "jquery": "3.3.1",
     "ng2-toasty": "^4.0.3",
