+++ ClientApp/src/app/components/admin/admin.component.html
@@ -1,1 +1,2 @@
 <h1>Admin</h1>
+<chart type="pie" [data]="data"></chart>
