+++ ClientApp/src/app/components/admin/admin.component.ts
@@ -6,6 +6,19 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./admin.component.css']
 })
 export class AdminComponent implements OnInit {
+  data = {
+    labels: ['BMW', 'Audi', 'Mazda'],
+    datasets: [
+      {
+        data: [5, 3, 1],
+        backgroundColor: [
+          "#ff6384",
+          "#36a2eb",
+          "#ffce56"
+        ]
+      }
+    ]
+  }
 
   constructor() { }
