+++ ClientApp/src/app/components/vehicle-form/vehicle-form.component.html
@@ -54,6 +54,8 @@
     <label for="contactEmail">Email</label>
     <input id="contactEmail" type="text" class="form-control" [(ngModel)]="vehicle.contact.email" name="contactEmail">
   </div>
-  <button class="btn btn-primary" [disabled]="!f.valid">Save</button>&nbsp;&nbsp;
-  <button *ngIf="vehicle.id" class="btn btn-danger" type="button" (click)="delete()">Delete</button>
+  <p>
+    <button class="btn btn-primary" [disabled]="!f.valid">Save</button>&nbsp;&nbsp;
+    <button *ngIf="vehicle.id" class="btn btn-danger" type="button" (click)="delete()">Delete</button>
+  </p>
 </form>
