+++ appsettings.json
@@ -1,7 +1,4 @@
 {
-  "ConnectionStrings": {
-    "Default": "Server=.;Database=vega;User Id=sa;Password=IamTan86"
-  },
   "PhotoSettings": {
     "MaxBytes": 41943040,
     "AcceptedFileTypes": [".jpg", ".jpeg", ".png"]
